creating the API Interface to centralize the code of the API request in a single file
