--- android-client/Espoo/app/src/main/java/com/espoo/android/api/SurveyService.kt
@@ -1,10 +1,0 @@
-package com.espoo.android.api
-
-import com.espoo.android.model.Survey
-import retrofit2.Call
-import retrofit2.http.GET
-
-interface SurveyService {
-    @GET("/api/v1/surveys")
-    fun getSurveys(): Call<List<Survey>>
-}
