+++ android-client/Espoo/app/src/main/java/com/espoo/android/view/LoginActivity.kt
@@ -7,7 +7,7 @@ import android.util.Log
 import android.view.View
 import android.widget.Toast
 import com.espoo.android.R
-import com.espoo.android.api.UserService
+import com.espoo.android.api.ApiService
 import com.espoo.android.helper.SessionManager
 import com.espoo.android.helper.SessionManager.PreferencesConstants.API_TOKEN
 import com.espoo.android.helper.SessionManager.PreferencesConstants.IS_LOGIN
@@ -17,16 +17,13 @@ import com.espoo.android.model.AuthData
 import com.espoo.android.model.User
 import com.espoo.android.model.UserLogin
 import kotlinx.android.synthetic.main.activity_login.*
-import okhttp3.OkHttpClient
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 
 class LoginActivity : AppCompatActivity() {
 
-    private lateinit var service: UserService
+    private lateinit var service: ApiService
     private lateinit var sessionManager : SessionManager
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -38,28 +35,7 @@ class LoginActivity : AppCompatActivity() {
             openMainActivity()
         }
 
-        /* This code is due to a backend implementation tha needs to insert the application/json headers
-        *  manually in each request */
-        val httpClient = OkHttpClient.Builder()
-        httpClient.addInterceptor { chain ->
-            val request = chain
-                .request()
-                .newBuilder()
-                .header("Accept", "application/json")
-                .header("Content-Type", "application/json")
-                .build()
-
-            chain.proceed(request)
-        }
-
-        val okHttpClient = httpClient.build()
-
-        service = Retrofit.Builder()
-            .baseUrl("https://espoo.herokuapp.com")
-            .client(okHttpClient)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(UserService::class.java)
+        service = ApiService.create("")
     }
 
     fun login(view: View) {
